+++ mackerel-plugin-aws-rekognition/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/mackerelio/mackerel-agent-plugins/mackerel-plugin-aws-rekognition/lib"
+
+func main() {
+	mpawsrekognition.Do()
+}